add docs-generated and GitHub desktop instructions
During recent training sessions with the docs team, we discovered that the current publishing instructions do not mention preparing the generated docs before the release and do not contain sufficient detail for someone who is working from GitHub Desktop and unfamiliar with the git CLI (which is true for many writers).

Rather than make the original instructions more dense, adding a new section that describes the same process using the desktop client. Let me know what you think. We can iterate as needed.